requirements.txt
PyYAML==5.3.1
spotipy==2.12.0
